+++ packages/avalanche-module/src/utils/transaction-detail-sections/subnet-detail-section.ts
@@ -1,16 +1,17 @@
 import type { CreateSubnetTx, DetailSection } from '@avalabs/vm-module-types';
 import { currencyItem, textItem } from '@internal/utils';
 import { AVAX_NONEVM_DENOMINATION } from '../../constants';
+import { addressListItem } from '@internal/utils/src/utils/detail-item';
 
 export const subnetDetailSection = (tx: CreateSubnetTx, symbol: string) => {
   const details: DetailSection[] = [];
   const { txFee, threshold, controlKeys } = tx;
 
   details.push({
-    title: 'Subnet Details',
+    title: 'L1 Details',
     items: [
-      textItem(controlKeys.length > 1 ? 'Owners' : 'Owner', controlKeys.join('\n'), 'vertical'),
-      textItem('Signature Threshold', `${threshold}/${controlKeys.length}`, 'vertical'),
+      addressListItem(controlKeys.length > 1 ? 'Owners' : 'Owner', controlKeys),
+      textItem('Signature Threshold', `${threshold}/${controlKeys.length}`, 'horizontal'),
     ],
   });
   if (txFee) {