+++ packages/evm-module/src/utils/get-provider.ts
@@ -1,6 +1,6 @@
 import { JsonRpcBatchInternal } from '@avalabs/core-wallets-sdk';
 import { getGlacierApiKey } from '@internal/utils';
-import { Network as EVMNetwork } from 'ethers';
+import { FetchRequest, Network as EVMNetwork } from 'ethers';
 
 type ProviderParams = {
   chainId: number;
@@ -8,14 +8,28 @@ type ProviderParams = {
   rpcUrl: string;
   multiContractAddress?: string;
   pollingInterval?: number;
+  customRpcHeaders?: Record<string, string>;
 };
 
 export const getProvider = async (params: ProviderParams): Promise<JsonRpcBatchInternal> => {
-  const { chainId, chainName, rpcUrl, multiContractAddress, pollingInterval = 2000 } = params;
+  const { chainId, chainName, rpcUrl, multiContractAddress, pollingInterval = 2000, customRpcHeaders } = params;
+
+  const url = addGlacierAPIKeyIfNeeded(rpcUrl);
+  let urlOrFetchRequest: string | FetchRequest = url;
+
+  if (customRpcHeaders) {
+    const fetchRequest = new FetchRequest(url);
+
+    for (const [name, value] of Object.entries(customRpcHeaders)) {
+      fetchRequest.setHeader(name, value);
+    }
+
+    urlOrFetchRequest = fetchRequest;
+  }
 
   const provider = new JsonRpcBatchInternal(
     { maxCalls: 40, multiContractAddress },
-    addGlacierAPIKeyIfNeeded(rpcUrl),
+    urlOrFetchRequest,
     new EVMNetwork(chainName, chainId),
   );